chore: give launch its own `LaunchApi` and drop `InternalApi` from it
Description
-----------

Launch was the largest `InternalApi` user: run queue and launch agent mutations, run state lookups, and the settings, API key and default entity it passed around as `api`. Those operations now live on `LaunchApi`, a `wandb.Api` subclass in `wandb.sdk.launch.api`, so launch code and its tests keep calling `api.pop_from_run_queue(...)` and friends while the object underneath is the public API. The launch sweep scheduler reaches the sweep helpers from the previous PR through that same object instead of a second `wandb.Api`, and the `wandb launch`, `launch-agent`, `launch-sweep`, `scheduler` and `job create` commands construct a `LaunchApi` directly. `wandb.Api()` verifies credentials on construction, so the separate `check_logged_in` step is gone.

`InternalApi` loses the run queue, launch agent, run state and sweep state methods. The `updateArtifact` mutation the job builder uses moves next to the other artifact mutations in `wandb.sdk.artifacts._gqlutils`.

- [x] I updated CHANGELOG.unreleased.md, or it's not applicable

Testing
-------

Launch unit tests; the agent tests keep their `MagicMock` API since the method names did not change.
