+++ tests/system_tests/test_launch/test_launch_cli.py
@@ -19,7 +19,7 @@ def _setup_agent(monkeypatch, pop_func):
     )
 
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.create_launch_agent",
+        "wandb.sdk.launch.api.LaunchApi.create_launch_agent",
         lambda c, e, p, q, a, v: {"launchAgentId": "mock_agent_id"},
     )
 
@@ -32,11 +32,11 @@ def test_agent_stop_polling(runner, monkeypatch, user, test_settings):
     _setup_agent(monkeypatch, patched_pop_empty_queue)
 
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.get_launch_agent",
+        "wandb.sdk.launch.api.LaunchApi.get_launch_agent",
         lambda c, i: {"id": "mock_agent_id", "name": "blah", "stopPolling": True},
     )
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.update_launch_agent_status",
+        "wandb.sdk.launch.api.LaunchApi.update_launch_agent_status",
         lambda c, i, s: {"success": True},
     )
 
@@ -61,11 +61,11 @@ def test_agent_update_failed(runner, monkeypatch, user, test_settings):
     _setup_agent(monkeypatch, patched_pop_empty_queue)
 
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.get_launch_agent",
+        "wandb.sdk.launch.api.LaunchApi.get_launch_agent",
         lambda c, i: {"id": "mock_agent_id", "name": "blah", "stopPolling": False},
     )
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.update_launch_agent_status",
+        "wandb.sdk.launch.api.LaunchApi.update_launch_agent_status",
         lambda c, i, s: {"success": False},
     )
 
@@ -98,12 +98,12 @@ def test_launch_agent_launch_error_continue(runner, monkeypatch, user, test_sett
     )
 
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.get_launch_agent",
+        "wandb.sdk.launch.api.LaunchApi.get_launch_agent",
         lambda c, i: {"id": "mock_agent_id", "name": "blah", "stopPolling": False},
     )
 
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.update_launch_agent_status",
+        "wandb.sdk.launch.api.LaunchApi.update_launch_agent_status",
         lambda c, i, s: {"success": True},
     )
 
@@ -218,11 +218,11 @@ def test_launch_supplied_logfile(runner, monkeypatch, wandb_caplog, user):
     _setup_agent(monkeypatch, patched_pop_empty_queue)
 
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.get_launch_agent",
+        "wandb.sdk.launch.api.LaunchApi.get_launch_agent",
         lambda c, i: {"id": "mock_agent_id", "name": "blah", "stopPolling": True},
     )
     monkeypatch.setattr(
-        "wandb.sdk.internal.internal_api.Api.update_launch_agent_status",
+        "wandb.sdk.launch.api.LaunchApi.update_launch_agent_status",
         lambda c, i, s: {"success": True},
     )
 
@@ -333,17 +333,12 @@ def test_launch_template_vars(command_inputs, expected_error, runner, monkeypatc
         mock_rq.name = "test-queue"
         return mock_rq
 
-    def patched_public_api(*args, **kwargs):
+    def patched_launch_api(*args, **kwargs):
         mock_api = Mock()
         mock_api.run_queue.side_effect = patched_run_queue
         return mock_api
 
-    monkeypatch.setattr(
-        "wandb.cli.cli.PublicApi",
-        patched_public_api,
-    )
-
-    monkeypatch.setattr("wandb.cli.cli.launch_utils.check_logged_in", lambda _: None)
+    monkeypatch.setattr("wandb.cli.cli.LaunchApi", patched_launch_api)
 
     result = "none"
     with runner.isolated_filesystem():
