+++ wandb/sdk/internal/internal_api.py
@@ -4,7 +4,6 @@ import json
 import logging
 import os
 import re
-import socket
 import tempfile
 from collections.abc import Mapping, MutableMapping
 from copy import deepcopy
@@ -12,17 +11,12 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, overload
 
 import wandb
-from wandb import env, util
+from wandb import env
 from wandb.analytics import TelemetryRecorder, get_telemetry_recorder
 from wandb.apis.normalize import normalize_exceptions
 from wandb.errors import AuthenticationError, CommError, UsageError
 from wandb.integration.sagemaker import parse_sm_secrets
-from wandb.proto.wandb_api_pb2 import (
-    ApiRequest,
-    CreateRunQueueRequest,
-    DownloadFileRequest,
-    RunQueueOperationRequest,
-)
+from wandb.proto.wandb_api_pb2 import ApiRequest, DownloadFileRequest
 from wandb.sdk import wandb_setup
 from wandb.sdk.lib.hashutil import B64Digest, md5_file_b64
 
@@ -31,10 +25,8 @@ from ..lib.filenames import DIFF_FNAME, METADATA_FNAME
 
 logger = logging.getLogger(__name__)
 
-LAUNCH_DEFAULT_PROJECT = "model-registry"
-
 if TYPE_CHECKING:
-    from typing import Literal, TypedDict
+    from typing import TypedDict
 
     from wandb.apis.public.service_api import ServiceApi
 
@@ -340,75 +332,6 @@ class Api:
         return project, run
 
     @normalize_exceptions
-    def fail_run_queue_item(
-        self,
-        run_queue_item_id: str,
-        message: str,
-        stage: str,
-        file_paths: list[str] | None = None,
-    ) -> bool:
-        variables: dict[str, str | (list[str] | None)] = {
-            "runQueueItemId": run_queue_item_id,
-            "message": message,
-            "stage": stage,
-        }
-        if file_paths is not None:
-            variables["filePaths"] = file_paths
-        mutation_string = """
-        mutation failRunQueueItem($runQueueItemId: ID!, $message: String!, $stage: String!, $filePaths: [String!]) {
-            failRunQueueItem(
-                input: {
-                    runQueueItemId: $runQueueItemId
-                    message: $message
-                    stage: $stage
-                    filePaths: $filePaths
-                }
-            ) {
-                success
-            }
-        }
-        """
-
-        mutation = mutation_string
-        response = self.execute(mutation, variables=variables)
-        result: bool = response["failRunQueueItem"]["success"]
-        return result
-
-    @normalize_exceptions
-    def update_run_queue_item_warning(
-        self,
-        run_queue_item_id: str,
-        message: str,
-        stage: str,
-        file_paths: list[str] | None = None,
-    ) -> bool:
-        mutation = """
-        mutation updateRunQueueItemWarning($runQueueItemId: ID!, $message: String!, $stage: String!, $filePaths: [String!]) {
-            updateRunQueueItemWarning(
-                input: {
-                    runQueueItemId: $runQueueItemId
-                    message: $message
-                    stage: $stage
-                    filePaths: $filePaths
-                }
-            ) {
-                success
-            }
-        }
-        """
-        response = self.execute(
-            mutation,
-            variables={
-                "runQueueItemId": run_queue_item_id,
-                "message": message,
-                "stage": stage,
-                "filePaths": file_paths,
-            },
-        )
-        result: bool = response["updateRunQueueItemWarning"]["success"]
-        return result
-
-    @normalize_exceptions
     def viewer(self) -> dict[str, Any]:
         query = """
         query Viewer{
@@ -691,39 +614,6 @@ class Api:
 
         return commit, config, patch, metadata
 
-    @normalize_exceptions
-    def check_stop_requested(
-        self, project_name: str, entity_name: str, run_id: str
-    ) -> bool:
-        query = """
-        query RunStoppedStatus($projectName: String, $entityName: String, $runId: String!) {
-            project(name:$projectName, entityName:$entityName) {
-                run(name:$runId) {
-                    stopped
-                }
-            }
-        }
-        """
-
-        response = self.execute(
-            query,
-            variables={
-                "projectName": project_name,
-                "entityName": entity_name,
-                "runId": run_id,
-            },
-        )
-
-        project = response.get("project", None)
-        if not project:
-            return False
-        run = project.get("run", None)
-        if not run:
-            return False
-
-        status: bool = run["stopped"]
-        return status
-
     def format_project(self, project: str) -> str:
         return re.sub(r"\W+", "-", project.lower()).strip("-_")
 
@@ -763,693 +653,6 @@ class Api:
         )
         result: dict[str, Any] = response["upsertModel"]["model"]
         return result
-
-    @normalize_exceptions
-    def entity_is_team(self, entity: str) -> bool:
-        query = """
-            query EntityIsTeam($entity: String!) {
-                entity(name: $entity) {
-                    id
-                    isTeam
-                }
-            }
-            """
-        variables = {
-            "entity": entity,
-        }
-
-        res = self.execute(query, variables)
-        if res.get("entity") is None:
-            raise Exception(
-                f"Error fetching entity {entity} "
-                "check that you have access to this entity"
-            )
-
-        is_team: bool = res["entity"]["isTeam"]
-        return is_team
-
-    @normalize_exceptions
-    def get_project_run_queues(self, entity: str, project: str) -> list[dict[str, str]]:
-        query = """
-        query ProjectRunQueues($entity: String!, $projectName: String!){
-            project(entityName: $entity, name: $projectName) {
-                runQueues {
-                    id
-                    name
-                    createdBy
-                    access
-                }
-            }
-        }
-        """
-        variables = {
-            "projectName": project,
-            "entity": entity,
-        }
-
-        res = self.execute(query, variables)
-        if res.get("project") is None:
-            # circular dependency: (LAUNCH_DEFAULT_PROJECT = model-registry)
-            if project == "model-registry":
-                msg = (
-                    f"Error fetching run queues for {entity} "
-                    "check that you have access to this entity and project"
-                )
-            else:
-                msg = (
-                    f"Error fetching run queues for {entity}/{project} "
-                    "check that you have access to this entity and project"
-                )
-
-            raise Exception(msg)
-
-        project_run_queues: list[dict[str, str]] = res["project"]["runQueues"]
-        return project_run_queues
-
-    @normalize_exceptions
-    def push_to_run_queue_by_name(
-        self,
-        entity: str,
-        project: str,
-        queue_name: str,
-        run_spec: str,
-        template_variables: dict[str, int | float | str] | None,
-        priority: int | None = None,
-    ) -> dict[str, Any] | None:
-        mutation_params = """
-            $entityName: String!,
-            $projectName: String!,
-            $queueName: String!,
-            $runSpec: JSONString!
-        """
-
-        mutation_input = """
-            entityName: $entityName,
-            projectName: $projectName,
-            queueName: $queueName,
-            runSpec: $runSpec
-        """
-
-        variables: dict[str, Any] = {
-            "entityName": entity,
-            "projectName": project,
-            "queueName": queue_name,
-            "runSpec": run_spec,
-        }
-        if priority is not None:
-            variables["priority"] = priority
-            mutation_params += ", $priority: Int"
-            mutation_input += ", priority: $priority"
-
-        if template_variables is not None:
-            variables.update({"templateVariableValues": json.dumps(template_variables)})
-            mutation_params += ", $templateVariableValues: JSONString"
-            mutation_input += ", templateVariableValues: $templateVariableValues"
-
-        mutation = f"""
-        mutation pushToRunQueueByName(
-          {mutation_params}
-        ) {{
-            pushToRunQueueByName(
-                input: {{
-                    {mutation_input}
-                }}
-            ) {{
-                runQueueItemId
-                runSpec
-            }}
-        }}
-        """
-
-        try:
-            result: dict[str, Any] | None = self.execute(mutation, variables).get(
-                "pushToRunQueueByName"
-            )
-            if not result:
-                return None
-
-            if result.get("runSpec"):
-                run_spec = json.loads(str(result["runSpec"]))
-                result["runSpec"] = run_spec
-
-            return result
-        except Exception as e:
-            if (
-                'Cannot query field "runSpec" on type "PushToRunQueueByNamePayload"'
-                not in str(e)
-            ):
-                return None
-
-        mutation_no_runspec = """
-        mutation pushToRunQueueByName(
-            $entityName: String!,
-            $projectName: String!,
-            $queueName: String!,
-            $runSpec: JSONString!,
-        ) {
-            pushToRunQueueByName(
-                input: {
-                    entityName: $entityName,
-                    projectName: $projectName,
-                    queueName: $queueName,
-                    runSpec: $runSpec
-                }
-            ) {
-                runQueueItemId
-            }
-        }
-        """
-
-        try:
-            result = self.execute(mutation_no_runspec, variables).get(
-                "pushToRunQueueByName"
-            )
-        except Exception:
-            result = None
-
-        return result
-
-    @normalize_exceptions
-    def push_to_run_queue(
-        self,
-        queue_name: str,
-        launch_spec: dict[str, str],
-        template_variables: dict | None,
-        project_queue: str,
-        priority: int | None = None,
-    ) -> dict[str, Any] | None:
-        entity = launch_spec.get("queue_entity") or launch_spec["entity"]
-        run_spec = json.dumps(launch_spec)
-
-        push_result = self.push_to_run_queue_by_name(
-            entity, project_queue, queue_name, run_spec, template_variables, priority
-        )
-
-        if push_result:
-            return push_result
-
-        if priority is not None:
-            # Cannot proceed with legacy method if priority is set
-            return None
-
-        """ Legacy Method """
-        queues_found = self.get_project_run_queues(entity, project_queue)
-        matching_queues = [
-            q
-            for q in queues_found
-            if q["name"] == queue_name
-            # ensure user has access to queue
-            and (
-                # TODO: User created queues in the UI have USER access
-                q["access"] in ["PROJECT", "USER"]
-                or q["createdBy"] == self.default_entity
-            )
-        ]
-        if not matching_queues:
-            # in the case of a missing default queue. create it
-            if queue_name == "default":
-                wandb.termlog(
-                    f"No default queue existing for entity: {entity} in project: {project_queue}, creating one."
-                )
-                create_queue_response = self._service_api.send_api_request(
-                    ApiRequest(
-                        run_queue_operation_request=RunQueueOperationRequest(
-                            create_run_queue_request=CreateRunQueueRequest(
-                                entity=launch_spec["entity"],
-                                project=project_queue,
-                                queue_name=queue_name,
-                                access="PROJECT",
-                            )
-                        )
-                    )
-                )
-                queue_result = create_queue_response.run_queue_operation_response.create_run_queue_response
-                if not queue_result.success or not queue_result.queue_id:
-                    wandb.termerror(
-                        f"Unable to create default queue for entity: {entity} on project: {project_queue}. Run could not be added to a queue"
-                    )
-                    return None
-                queue_id = queue_result.queue_id
-
-            else:
-                if project_queue == "model-registry":
-                    _msg = f"Unable to push to run queue {queue_name}. Queue not found."
-                else:
-                    _msg = f"Unable to push to run queue {project_queue}/{queue_name}. Queue not found."
-                wandb.termwarn(_msg)
-                return None
-        elif len(matching_queues) > 1:
-            wandb.termerror(
-                f"Unable to push to run queue {queue_name}. More than one queue found with this name."
-            )
-            return None
-        else:
-            queue_id = matching_queues[0]["id"]
-        spec_json = json.dumps(launch_spec)
-        variables = {"queueID": queue_id, "runSpec": spec_json}
-
-        mutation_params = """
-            $queueID: ID!,
-            $runSpec: JSONString!
-        """
-        mutation_input = """
-            queueID: $queueID,
-            runSpec: $runSpec
-        """
-        if template_variables is not None:
-            mutation_params += ", $templateVariableValues: JSONString"
-            mutation_input += ", templateVariableValues: $templateVariableValues"
-            variables.update({"templateVariableValues": json.dumps(template_variables)})
-
-        mutation = f"""
-        mutation pushToRunQueue(
-            {mutation_params}
-            ) {{
-            pushToRunQueue(
-                input: {{{mutation_input}}}
-            ) {{
-                runQueueItemId
-            }}
-        }}
-        """
-
-        response = self.execute(mutation, variables=variables)
-        if not response.get("pushToRunQueue"):
-            raise CommError(f"Error pushing run queue item to queue {queue_name}.")
-
-        result: dict[str, Any] | None = response["pushToRunQueue"]
-        return result
-
-    @normalize_exceptions
-    def pop_from_run_queue(
-        self,
-        queue_name: str,
-        entity: str | None = None,
-        project: str | None = None,
-        agent_id: str | None = None,
-    ) -> dict[str, Any] | None:
-        mutation = """
-        mutation popFromRunQueue($entity: String!, $project: String!, $queueName: String!, $launchAgentId: ID)  {
-            popFromRunQueue(input: {
-                entityName: $entity,
-                projectName: $project,
-                queueName: $queueName,
-                launchAgentId: $launchAgentId
-            }) {
-                runQueueItemId
-                runSpec
-            }
-        }
-        """
-        response = self.execute(
-            mutation,
-            variables={
-                "entity": entity,
-                "project": project,
-                "queueName": queue_name,
-                "launchAgentId": agent_id,
-            },
-        )
-        result: dict[str, Any] | None = response["popFromRunQueue"]
-        return result
-
-    @normalize_exceptions
-    def ack_run_queue_item(self, item_id: str, run_id: str | None = None) -> bool:
-        mutation = """
-        mutation ackRunQueueItem($itemId: ID!, $runId: String!)  {
-            ackRunQueueItem(input: { runQueueItemId: $itemId, runName: $runId }) {
-                success
-            }
-        }
-        """
-        response = self.execute(
-            mutation, variables={"itemId": item_id, "runId": str(run_id)}
-        )
-        if not response["ackRunQueueItem"]["success"]:
-            raise CommError(
-                "Error acking run queue item. Item may have already been acknowledged by another process"
-            )
-        result: bool = response["ackRunQueueItem"]["success"]
-        return result
-
-    @normalize_exceptions
-    def create_launch_agent(
-        self,
-        entity: str,
-        project: str,
-        queues: list[str],
-        agent_config: dict[str, Any],
-        version: str,
-    ) -> dict:
-        project_queues = self.get_project_run_queues(entity, project)
-        if not project_queues:
-            # create default queue if it doesn't already exist
-            response = self._service_api.send_api_request(
-                ApiRequest(
-                    run_queue_operation_request=RunQueueOperationRequest(
-                        create_run_queue_request=CreateRunQueueRequest(
-                            entity=entity,
-                            project=project,
-                            queue_name="default",
-                            access="PROJECT",
-                        )
-                    )
-                )
-            )
-            default = response.run_queue_operation_response.create_run_queue_response
-            if not default.success or not default.queue_id:
-                raise CommError(
-                    f"Unable to create default queue for {entity}/{project}. No queues for agent to poll"
-                )
-            project_queues = [{"id": default.queue_id, "name": "default"}]
-        polling_queue_ids = [
-            q["id"] for q in project_queues if q["name"] in queues
-        ]  # filter to poll specified queues
-        if len(polling_queue_ids) != len(queues):
-            raise CommError(
-                f"Could not start launch agent: Not all of requested queues ({', '.join(queues)}) found. "
-                f"Available queues for this project: {','.join([q['name'] for q in project_queues])}"
-            )
-
-        hostname = socket.gethostname()
-
-        variables = {
-            "entity": entity,
-            "project": project,
-            "queues": polling_queue_ids,
-            "hostname": hostname,
-            "agentConfig": json.dumps(agent_config),
-            "version": version,
-        }
-
-        mutation_params = """
-            $entity: String!,
-            $project: String!,
-            $queues: [ID!]!,
-            $hostname: String!,
-            $agentConfig: JSONString,
-            $version: String
-        """
-
-        mutation_input = """
-            entityName: $entity,
-            projectName: $project,
-            runQueues: $queues,
-            hostname: $hostname,
-            agentConfig: $agentConfig,
-            version: $version
-        """
-
-        mutation = f"""
-            mutation createLaunchAgent(
-                {mutation_params}
-            ) {{
-                createLaunchAgent(
-                    input: {{
-                        {mutation_input}
-                    }}
-                ) {{
-                    launchAgentId
-                }}
-            }}
-            """
-        result: dict = self.execute(mutation, variables)["createLaunchAgent"]
-        return result
-
-    @normalize_exceptions
-    def update_launch_agent_status(
-        self,
-        agent_id: str,
-        status: str,
-    ) -> dict:
-        mutation = """
-            mutation updateLaunchAgent($agentId: ID!, $agentStatus: String){
-                updateLaunchAgent(
-                    input: {
-                        launchAgentId: $agentId
-                        agentStatus: $agentStatus
-                    }
-                ) {
-                    success
-                }
-            }
-            """
-        variables = {
-            "agentId": agent_id,
-            "agentStatus": status,
-        }
-        result: dict = self.execute(mutation, variables)["updateLaunchAgent"]
-        return result
-
-    @normalize_exceptions
-    def get_launch_agent(self, agent_id: str) -> dict:
-        query = """
-            query LaunchAgent($agentId: ID!) {
-                launchAgent(id: $agentId) {
-                    id
-                    name
-                    runQueues
-                    hostname
-                    agentStatus
-                    stopPolling
-                    heartbeatAt
-                }
-            }
-            """
-        variables = {
-            "agentId": agent_id,
-        }
-        result: dict = self.execute(query, variables)["launchAgent"]
-        return result
-
-    @normalize_exceptions
-    def upsert_run(
-        self,
-        id: str | None = None,
-        name: str | None = None,
-        project: str | None = None,
-        host: str | None = None,
-        group: str | None = None,
-        tags: list[str] | None = None,
-        config: dict | None = None,
-        description: str | None = None,
-        entity: str | None = None,
-        state: str | None = None,
-        display_name: str | None = None,
-        notes: str | None = None,
-        repo: str | None = None,
-        job_type: str | None = None,
-        program_path: str | None = None,
-        commit: str | None = None,
-        sweep_name: str | None = None,
-        summary_metrics: str | None = None,
-        num_retries: int | None = None,
-    ) -> tuple[dict, bool]:
-        """Update a run.
-
-        Args:
-            id (str, optional): The existing run to update
-            name (str, optional): The name of the run to create
-            group (str, optional): Name of the group this run is a part of
-            project (str, optional): The name of the project
-            host (str, optional): The name of the host
-            tags (list, optional): A list of tags to apply to the run
-            config (dict, optional): The latest config params
-            description (str, optional): A description of this project
-            entity (str, optional): The entity to scope this project to.
-            display_name (str, optional): The display name of this project
-            notes (str, optional): Notes about this run
-            repo (str, optional): Url of the program's repository.
-            state (str, optional): State of the program.
-            job_type (str, optional): Type of job, e.g 'train'.
-            program_path (str, optional): Path to the program.
-            commit (str, optional): The Git SHA to associate the run with
-            sweep_name (str, optional): The name of the sweep this run is a part of
-            summary_metrics (str, optional): The JSON summary metrics
-            num_retries (int, optional): Number of retries
-        """
-        query_string = """
-        mutation UpsertBucket(
-            $id: String,
-            $name: String,
-            $project: String,
-            $entity: String,
-            $groupName: String,
-            $description: String,
-            $displayName: String,
-            $notes: String,
-            $commit: String,
-            $config: JSONString,
-            $host: String,
-            $debug: Boolean,
-            $program: String,
-            $repo: String,
-            $jobType: String,
-            $state: String,
-            $sweep: String,
-            $tags: [String!],
-            $summaryMetrics: JSONString,
-        ) {
-            upsertBucket(input: {
-                id: $id,
-                name: $name,
-                groupName: $groupName,
-                modelName: $project,
-                entityName: $entity,
-                description: $description,
-                displayName: $displayName,
-                notes: $notes,
-                config: $config,
-                commit: $commit,
-                host: $host,
-                debug: $debug,
-                jobProgram: $program,
-                jobRepo: $repo,
-                jobType: $jobType,
-                state: $state,
-                sweep: $sweep,
-                tags: $tags,
-                summaryMetrics: $summaryMetrics,
-            }) {
-                bucket {
-                    id
-                    name
-                    displayName
-                    description
-                    config
-                    sweepName
-                    project {
-                        id
-                        name
-                        entity {
-                            id
-                            name
-                        }
-                    }
-                    historyLineCount
-                }
-                inserted
-            }
-        }
-        """
-
-        mutation = query_string
-        config_str = json.dumps(config) if config else None
-        if not description or description.isspace():
-            description = None
-
-        variables = {
-            "id": id,
-            "entity": entity or self.settings("entity"),
-            "name": name,
-            "project": project or util.auto_project_name(program_path),
-            "groupName": group,
-            "tags": tags,
-            "description": description,
-            "config": config_str,
-            "commit": commit,
-            "displayName": display_name,
-            "notes": notes,
-            "host": None
-            if self.settings().get("anonymous") in ["allow", "must"]
-            else host,
-            "debug": env.is_debug(env=self._environ),
-            "repo": repo,
-            "program": program_path,
-            "jobType": job_type,
-            "state": state,
-            "sweep": sweep_name,
-            "summaryMetrics": summary_metrics,
-        }
-
-        response = self.execute(
-            mutation,
-            variables=variables,
-        )
-
-        run_obj: dict[str, dict[str, dict[str, str]]] = response["upsertBucket"][
-            "bucket"
-        ]
-        project_obj: dict[str, dict[str, str]] = run_obj.get("project", {})
-        if project_obj:
-            self.set_setting("project", project_obj["name"])
-            entity_obj = project_obj.get("entity", {})
-            if entity_obj:
-                self.set_setting("entity", entity_obj["name"])
-
-        return (
-            response["upsertBucket"]["bucket"],
-            response["upsertBucket"]["inserted"],
-        )
-
-    @normalize_exceptions
-    def get_run_info(
-        self,
-        entity: str,
-        project: str,
-        name: str,
-    ) -> dict:
-        query = """
-        query RunInfo($project: String!, $entity: String!, $name: String!) {
-            project(name: $project, entityName: $entity) {
-                run(name: $name) {
-                    runInfo {
-                        program
-                        args
-                        os
-                        python
-                        colab
-                        executable
-                        codeSaved
-                        cpuCount
-                        gpuCount
-                        gpu
-                        git {
-                            remote
-                            commit
-                        }
-                    }
-                }
-            }
-        }
-        """
-        variables = {"project": project, "entity": entity, "name": name}
-        res = self.execute(query, variables)
-        if res.get("project") is None:
-            raise CommError(
-                f"Error fetching run info for {entity}/{project}/{name}. Check that this project exists and you have access to this entity and project"
-            )
-        elif res["project"].get("run") is None:
-            raise CommError(
-                f"Error fetching run info for {entity}/{project}/{name}. Check that this run id exists"
-            )
-        run_info: dict = res["project"]["run"]["runInfo"]
-        return run_info
-
-    @normalize_exceptions
-    def get_run_state(self, entity: str, project: str, name: str) -> str:
-        query = """
-        query RunState(
-            $project: String!,
-            $entity: String!,
-            $name: String!) {
-            project(name: $project, entityName: $entity) {
-                run(name: $name) {
-                    state
-                }
-            }
-        }
-        """
-        variables = {
-            "project": project,
-            "entity": entity,
-            "name": name,
-        }
-        res = self.execute(query, variables)
-        if res.get("project") is None or res["project"].get("run") is None:
-            raise CommError(f"Error fetching run state for {entity}/{project}/{name}.")
-        run_state: str = res["project"]["run"]["state"]
-        return run_state
 
     @normalize_exceptions
     def download_urls(
@@ -1787,28 +990,6 @@ class Api:
         )
         return response["updateArtifact"]["artifact"]
 
-    def get_sweep_state(
-        self, sweep: str, entity: str | None = None, project: str | None = None
-    ) -> SweepState:
-        query = """
-            query GetSweepState($entity: String, $project: String, $sweep: String!) {
-                project(name: $project, entityName: $entity) {
-                    sweep(sweepName: $sweep) {
-                        state
-                    }
-                }
-            }
-            """
-        response = self.execute(
-            query,
-            variables={
-                "sweep": sweep,
-                "entity": entity or self.settings("entity"),
-                "project": project or self.settings("project"),
-            },
-        )
-        return response["project"]["sweep"]["state"]
-
     def set_sweep_state(
         self,
         sweep: str,
@@ -1856,30 +1037,3 @@ class Api:
     def _flatten_edges(self, response: _Response) -> list[dict]:
         """Return an array from the nested graphql relay structure."""
         return [node["node"] for node in response["edges"]]
-
-    @normalize_exceptions
-    def stop_run(
-        self,
-        run_id: str,
-    ) -> bool:
-        mutation = """
-            mutation stopRun($id: ID!) {
-                stopRun(input: {
-                    id: $id
-                }) {
-                    clientMutationId
-                    success
-                }
-            }
-            """
-
-        response = self.execute(
-            mutation,
-            variables={
-                "id": run_id,
-            },
-        )
-
-        success: bool = response["stopRun"].get("success")
-
-        return success
