+++ wandb/sdk/internal/job_builder.py
@@ -11,16 +11,17 @@ from collections.abc import Callable
 from typing import TYPE_CHECKING, Any, Literal, TypedDict
 
 import wandb
+from wandb.sdk.artifacts._gqlutils import update_artifact_metadata
 from wandb.sdk.artifacts._internal_artifact import InternalArtifact
 from wandb.sdk.artifacts.artifact import Artifact
 from wandb.sdk.data_types._dtypes import TypeRegistry
-from wandb.sdk.internal.internal_api import Api
 from wandb.sdk.lib.filenames import DIFF_FNAME, METADATA_FNAME, REQUIREMENTS_FNAME
 from wandb.util import make_artifact_name_safe
 
 _logger = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
+    from wandb.apis.public import Api
     from wandb.proto.wandb_internal_pb2 import ArtifactRecord
 
 FROZEN_REQUIREMENTS_FNAME = "requirements.frozen.txt"
@@ -455,9 +456,8 @@ class JobBuilder:
                 "input_types": {"@wandb.config": self.input_types},
                 "output_types": self.output_types,
             }
-            api.update_artifact_metadata(
-                self._partial_source_id,
-                new_metadata,
+            update_artifact_metadata(
+                api._service_api, self._partial_source_id, new_metadata
             )
             return None
