+++ wandb/sdk/launch/sweeps/scheduler.py
@@ -41,7 +41,7 @@ from wandb.sdk.sweeps import RunState as RunState
 if TYPE_CHECKING:
     import wandb.apis.public as public
     from wandb.apis.public import QueuedRun, Run
-    from wandb.sdk.internal.internal_api import Api
+    from wandb.sdk.launch.api import LaunchApi
 
 
 _logger = logging.getLogger(__name__)
@@ -86,7 +86,7 @@ class Scheduler(ABC):
 
     def __init__(
         self,
-        api: Api,
+        api: LaunchApi,
         *args: Any | None,
         polling_sleep: float | None = None,
         sweep_id: str | None = None,
@@ -98,19 +98,17 @@ class Scheduler(ABC):
     ):
         import yaml
 
-        from wandb.apis.public import Api as PublicApi
         from wandb.apis.public.sweeps import _sweep_with_runs
 
         self._api = api
-        self._public_api = PublicApi()
         self._entity = (
             entity
             or os.environ.get("WANDB_ENTITY")
-            or api.settings("entity")
+            or api.settings["entity"]
             or api.default_entity
         )
         self._project = (
-            project or os.environ.get("WANDB_PROJECT") or api.settings("project")
+            project or os.environ.get("WANDB_PROJECT") or api.settings["project"]
         )
         self._sweep_id: str = sweep_id or "empty-sweep-id"
         self._state: SchedulerState = SchedulerState.PENDING
@@ -118,7 +116,7 @@ class Scheduler(ABC):
         # Make sure the provided sweep_id corresponds to a valid sweep
         try:
             resp = _sweep_with_runs(
-                self._public_api,
+                self._api,
                 sweep_id,
                 "{}",
                 entity=self._entity,
@@ -398,7 +396,7 @@ class Scheduler(ABC):
         """
         if self._kwargs.get("job"):
             try:
-                _job_artifact = self._public_api.job(self._kwargs["job"])
+                _job_artifact = self._api.job(self._kwargs["job"])
                 wandb.termlog(
                     f"{LOG_PREFIX}Successfully loaded job ({_job_artifact.name}) in scheduler"
                 )
@@ -418,7 +416,7 @@ class Scheduler(ABC):
             _logger.debug(f"{LOG_PREFIX}Starting AgentHeartbeat worker ({worker_id})")
             try:
                 worker = register_agent(
-                    self._public_api,
+                    self._api,
                     f"{socket.gethostname()}-{worker_id}",  # host
                     sweep_id=self._sweep_id,
                     project=self._project,
@@ -513,7 +511,7 @@ class Scheduler(ABC):
 
         try:
             sweep_state = _get_sweep_state(
-                self._public_api,
+                self._api,
                 self._sweep_id,
                 entity=self._entity,
                 project=self._project,
@@ -560,7 +558,7 @@ class Scheduler(ABC):
             if not queued_run:
                 return []
 
-            api_run: Run = self._public_api.run(
+            api_run: Run = self._api.run(
                 f"{queued_run.entity}/{queued_run.project}/{run_id}"
             )
             metric_name = self._sweep_config["metric"]["name"]
@@ -571,18 +569,6 @@ class Scheduler(ABC):
         except Exception as e:
             _logger.debug(f"[_get_metrics_from_run] {e}")
         return []
-
-    def _get_run_info(self, run_id: str) -> dict[str, Any]:
-        """Use the public api to get info about a run."""
-        try:
-            info: dict[str, Any] = self._api.get_run_info(
-                self._entity, self._project, run_id
-            )
-            if info:
-                return info
-        except Exception as e:
-            _logger.debug(f"[_get_run_info] {e}")
-        return {}
 
     def _get_run_state(
         self, run_id: str, prev_run_state: RunState = RunState.UNKNOWN
@@ -609,28 +595,13 @@ class Scheduler(ABC):
             run_state = RunState.UNKNOWN
         return run_state
 
-    def _create_run(self) -> dict[str, Any]:
-        """Use the public api to create a blank run."""
-        try:
-            server_run, inserted = self._api.upsert_run(
-                project=self._project,
-                entity=self._entity,
-                sweep_name=self._sweep_id,
-            )
-            return server_run
-        except Exception as e:
-            _logger.debug(f"[_create_run] {e}")
-            raise SchedulerError(
-                "Error creating run from scheduler, check API connection and CLI version."
-            )
-
     def _set_sweep_state(self, state: str) -> None:
         from wandb.apis.public.sweeps import _set_sweep_state
 
         wandb.termlog(f"{LOG_PREFIX}Updating sweep state to: {state.lower()}")
         try:
             _set_sweep_state(
-                self._public_api,
+                self._api,
                 self._sweep_id,
                 state,
                 entity=self._entity,
